Support for building C-Kermit for OS/2 with OpenWatcom 1.9

+++ kermit/k95/kui/kclient.cxx
@@ -30,7 +30,7 @@ extern int tt_cursor_blink;
 extern int tt_scrsize[];	/* Scrollback buffer size */
 extern int tt_status[];
 extern int scrollflag[];
-extern int vmode;
+extern BYTE vmode;
 extern int win32ScrollUp, win32ScrollDown;
 extern int trueblink, trueunderline, trueitalic;
 unsigned char geterasecolor(int);
